refactor: rename protobuf to proto3
