+++ src/noir/codec/CMakeLists.txt
@@ -1,3 +1,3 @@
-add_noir_test(protobuf_test test/protobuf_test.cpp)
-add_noir_test(rlp_test test/rlp_test.cpp)
-add_noir_test(scale_test test/scale_test.cpp)
+add_noir_test(proto3_test test/proto3_test.cpp DEPENDS noir_common)
+add_noir_test(rlp_test test/rlp_test.cpp DEPENDS noir_common)
+add_noir_test(scale_test test/scale_test.cpp DEPENDS noir_common)
